MCD-Agencia-main/backend/requirements.txt
# =============================================================================
# MCD-Agencia Backend - Python Dependencies
# =============================================================================
# Django & DRF Core
# =============================================================================
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# =============================================================================
# Authentication & Security
# =============================================================================
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=0.60,<1.0
dj-rest-auth>=5.0,<6.0
PyJWT>=2.8,<3.0
cryptography>=42.0,<44.0
python-dotenv>=1.0,<2.0

# =============================================================================
# Database
# =============================================================================
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0
django-redis>=5.4,<6.0

# =============================================================================
# Task Queue (Celery)
# =============================================================================
celery>=5.3,<6.0
redis>=5.0,<6.0

# =============================================================================
# File Storage (S3 Compatible)
# =============================================================================
boto3>=1.34,<2.0
django-storages>=1.14,<2.0

# =============================================================================
# PDF Generation
# =============================================================================
weasyprint>=60.0,<61.0
reportlab>=4.0,<5.0
openpyxl>=3.1,<4.0

# =============================================================================
# Email
# =============================================================================
django-anymail>=10.2,<11.0

# =============================================================================
# API Documentation
# =============================================================================
drf-spectacular>=0.27,<1.0

# =============================================================================
# Utilities
# =============================================================================
Pillow>=10.2,<11.0
python-slugify>=8.0,<9.0
django-mptt>=0.16,<1.0
django-extensions>=3.2,<4.0
django-import-export>=3.3,<4.0

# =============================================================================
# AI / Chatbot
# =============================================================================
google-genai>=1.0,<2.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
black>=24.1,<25.0
flake8>=7.0,<8.0
isort>=5.13,<6.0
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0

# =============================================================================
# Production
# =============================================================================
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
sentry-sdk>=1.39,<2.0
python-json-logger>=2.0,<3.0
django-celery-beat>=2.5,<3.0
